Move Wrapper component into a 'generic' folder
